+++ src/main/java/lotto/controller/LottoController.java
@@ -3,11 +3,11 @@ package lotto.controller;
 import camp.nextstep.edu.missionutils.Console;
 import java.util.List;
 import lotto.LottoConfig;
-import lotto.controller.dto.PurchaseRequest;
-import lotto.controller.dto.ResultRequest;
-import lotto.controller.dto.ResultResponse;
+import lotto.dto.PurchaseRequest;
+import lotto.dto.ResultRequest;
+import lotto.dto.ResultResponse;
 import lotto.domain.PlayRule;
-import lotto.controller.dto.PurchaseResponse;
+import lotto.dto.PurchaseResponse;
 import lotto.service.LottoService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
