+++ src/main/java/lotto/dto/PurchaseRequest.java
@@ -1,4 +1,4 @@
-package lotto.controller.dto;
+package lotto.dto;
 
 public record PurchaseRequest(
         String purchaseAmount
