+++ src/main/java/lotto/dto/ResultRequest.java
@@ -1,4 +1,4 @@
-package lotto.controller.dto;
+package lotto.dto;
 
 public record ResultRequest(
         String winningNumber,
